moved generic components into separate file and fixed bugs
